Zex/backend/requirements-production.txt
# Production requirements for AI SaaS Platform
# Core FastAPI and async support
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
authlib==1.3.0
itsdangerous==2.1.2

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.26.0

# Rate limiting & Caching
slowapi==0.1.9
redis==5.0.1

# Background tasks
celery==5.3.4

# Logging & Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0

# File processing
Pillow==12.0.0
python-magic==0.4.27
aiofiles==23.2.1

# Cloud storage
boto3==1.34.0
supabase==2.3.4

# System monitoring
psutil==5.9.8

# Web3 integration
web3==6.15.0
eth-account==0.11.0

# AI Provider SDKs
anthropic==0.18.1
google-cloud-texttospeech==2.16.3

# Additional utilities
python-dateutil==2.8.2
python-multipart==0.0.6
websockets==12.0